move cache implementation to a different file
